MOve azure-env
