+++ src/client/azure-environment/azure-environment.ts
