+++ src/client/azure-environment/index.ts
